+++ poem/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.3.6] 2022-2-22
+
+- Change charset from `utf8` to `utf-8`. [#213](https://github.com/poem-web/poem/pull/213)
+
 # [1.3.5] 2022-2-22
 
 - Add `X-Accel-Buffering: no` header to SSE response.